pyproject.toml
[project]
name = "ingest-llm-as"
version = "0.1.0"
description = "A microservice for ingesting data into the ApexSigma ecosystem."
license = "MIT"
authors = [
    {name = "SigmaDev11",email = "steynsean11@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
]

[tool.poetry]
packages = [{include = "ingest_llm_as", from = "src"}]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
pydantic-settings = "^2.3.4"
prometheus-client = "^0.19.0"
prometheus-fastapi-instrumentator = "^6.1.0"
opentelemetry-api = "^1.20.0"
opentelemetry-sdk = "^1.20.0"
opentelemetry-instrumentation-fastapi = "^0.41b0"
opentelemetry-exporter-jaeger-thrift = "^1.20.0"
opentelemetry-instrumentation-httpx = "^0.41b0"
structlog = "^23.1.0"
langfuse = "^2.14.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"
mypy = "^1.10.0"
pytest = "^8.2.2"
httpx = "^0.27.0"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
